Bring back testing utilities used in downstream packages
Follow up to
* 534a9ae28a6bfea3bc31921cd92e4e4d2875b6c9
* e1d36cacb2cfb946f56872ee03e8678dc6ccf7f4
