+++ pymc/testing.py
@@ -14,29 +14,49 @@
 import functools as ft
 import itertools as it
 
-from contextlib import ExitStack as does_not_raise
 from typing import Callable, List, Optional
 
 import numpy as np
-import numpy.random as nr
-import numpy.testing as npt
 import pytensor
-import pytensor.tensor as at
+import pytensor.tensor as pt
 import pytest
-import scipy.special as sp
-import scipy.stats as st
 
+from numpy import random as nr
+from numpy import testing as npt
 from pytensor.compile.mode import Mode
+from pytensor.graph.basic import ancestors
+from pytensor.graph.rewriting.basic import in2out
+from pytensor.tensor.random.op import RandomVariable
+from scipy import special as sp
+from scipy import stats as st
 
 import pymc as pm
 
+from pymc import logcdf, logp
 from pymc.distributions.shape_utils import change_dist_size
 from pymc.initial_point import make_initial_point_fn
-from pymc.logprob.abstract import logcdf
-from pymc.logprob.joint_logprob import joint_logp, logp
+from pymc.logprob import joint_logp
 from pymc.logprob.utils import ParameterValueError
-from pymc.pytensorf import compile_pymc, floatX, intX
-from tests.helpers import SeededTest, select_by_precision
+from pymc.pytensorf import (
+    compile_pymc,
+    floatX,
+    intX,
+    local_check_parameter_to_ninf_switch,
+)
+
+# This mode can be used for tests where model compilations takes the bulk of the runtime
+# AND where we don't care about posterior numerical or sampling stability (e.g., when
+# all that matters are the shape of the draws or deterministic values of observed data).
+# DO NOT USE UNLESS YOU HAVE A GOOD REASON TO!
+fast_unstable_sampling_mode = (
+    pytensor.compile.mode.FAST_COMPILE
+    # Remove slow rewrite phases
+    .excluding("canonicalize", "specialize")
+    # Include necessary rewrites for proper logp handling
+    .including("remove_TransformedVariables").register(
+        (in2out(local_check_parameter_to_ninf_switch), -1)
+    )
+)
 
 
 def product(domains, n_samples=-1):
@@ -47,7 +67,8 @@ def product(domains, n_samples=-1):
                  must be "domain-like", as in, have a `.vals` property
         n_samples: int, maximum samples to return.  -1 to return whole product
 
-    Returns:
+    Returns
+    -------
         list of the cartesian product of the domains
     """
     try:
@@ -119,22 +140,6 @@ class Domain:
         return Domain([-v for v in self.vals], self.dtype, (-self.lower, -self.upper), self.shape)
 
 
-@pytest.mark.parametrize(
-    "values, edges, expectation",
-    [
-        ([], None, pytest.raises(IndexError)),
-        ([], (0, 0), pytest.raises(ValueError)),
-        ([0], None, pytest.raises(ValueError)),
-        ([0], (0, 0), does_not_raise()),
-        ([-1, 1], None, pytest.raises(ValueError)),
-        ([-1, 0, 1], None, does_not_raise()),
-    ],
-)
-def test_domain(values, edges, expectation):
-    with expectation:
-        Domain(values, edges=edges)
-
-
 class ProductDomain:
     def __init__(self, domains):
         self.vals = list(it.product(*(d.vals for d in domains)))
@@ -203,22 +208,23 @@ Rplus = Domain([0, 0.01, 0.1, 0.9, 0.99, 1, 1.5, 2, 100, np.inf])
 Rplusbig = Domain([0, 0.5, 0.9, 0.99, 1, 1.5, 2, 20, np.inf])
 Rminusbig = Domain([-np.inf, -2, -1.5, -1, -0.99, -0.9, -0.5, -0.01, 0])
 Unit = Domain([0, 0.001, 0.1, 0.5, 0.75, 0.99, 1])
-
 Circ = Domain([-np.pi, -2.1, -1, -0.01, 0.0, 0.01, 1, 2.1, np.pi])
-
 Runif = Domain([-np.inf, -0.4, 0, 0.4, np.inf])
 Rdunif = Domain([-np.inf, -1, 0, 1, np.inf], "int64")
 Rplusunif = Domain([0, 0.5, np.inf])
 Rplusdunif = Domain([0, 10, np.inf], "int64")
-
 I = Domain([-np.inf, -3, -2, -1, 0, 1, 2, 3, np.inf], "int64")
-
 NatSmall = Domain([0, 3, 4, 5, np.inf], "int64")
 Nat = Domain([0, 1, 2, 3, np.inf], "int64")
 NatBig = Domain([0, 1, 2, 3, 5000, np.inf], "int64")
 PosNat = Domain([1, 2, 3, np.inf], "int64")
-
 Bool = Domain([0, 0, 1, 1], "int64")
+
+
+def select_by_precision(float64, float32):
+    """Helper function to choose reasonable decimal cutoffs for different floatX modes."""
+    decimal = float64 if pytensor.config.floatX == "float64" else float32
+    return decimal
 
 
 def build_model(distfam, valuedomain, vardomains, extra_args=None):
@@ -228,9 +234,9 @@ def build_model(distfam, valuedomain, vardomains, extra_args=None):
     with pm.Model() as m:
         param_vars = {}
         for v, dom in vardomains.items():
-            v_at = pytensor.shared(np.asarray(dom.vals[0]))
-            v_at.name = v
-            param_vars[v] = v_at
+            v_pt = pytensor.shared(np.asarray(dom.vals[0]))
+            v_pt.name = v
+            param_vars[v] = v_pt
         param_vars.update(extra_args)
         distfam(
             "value",
@@ -295,10 +301,10 @@ def check_logp(
         args.update(scipy_args)
         return scipy_logp(**args)
 
-    def _model_input_dict(model, param_vars, pt):
+    def _model_input_dict(model, param_vars, point):
         """Create a dict with only the necessary, transformed logp inputs."""
         pt_d = {}
-        for k, v in pt.items():
+        for k, v in point.items():
             rv_var = model.named_vars.get(k)
             nv = param_vars.get(k, rv_var)
             nv = model.rvs_to_values.get(nv, nv)
@@ -325,16 +331,16 @@ def check_logp(
     # Test supported value and parameters domain matches scipy
     domains = paramdomains.copy()
     domains["value"] = domain
-    for pt in product(domains, n_samples=n_samples):
-        pt = dict(pt)
-        pt_d = _model_input_dict(model, param_vars, pt)
+    for point in product(domains, n_samples=n_samples):
+        point = dict(point)
+        pt_d = _model_input_dict(model, param_vars, point)
         pt_logp = pm.Point(pt_d, model=model)
-        pt_ref = pm.Point(pt, filter_model_vars=False, model=model)
+        pt_ref = pm.Point(point, filter_model_vars=False, model=model)
         npt.assert_almost_equal(
             logp_pymc(pt_logp),
             logp_reference(pt_ref),
             decimal=decimal,
-            err_msg=str(pt),
+            err_msg=str(point),
         )
 
     valid_value = domain.vals[0]
@@ -360,7 +366,7 @@ def check_logp(
                 if invalid_edge is None:
                     continue
                 test_params = valid_params.copy()  # Shallow copy should be okay
-                test_params[invalid_param] = at.as_tensor_variable(invalid_edge)
+                test_params[invalid_param] = pt.as_tensor_variable(invalid_edge)
                 # We need to remove `Assert`s introduced by checks like
                 # `assert_negative_support` and disable test values;
                 # otherwise, we won't be able to create the `RandomVariable`
@@ -441,9 +447,6 @@ def check_logcdf(
         Whether to run test 2., which checks that pymc distribution logcdf
         returns -inf for invalid parameter values outside the supported domain edge
 
-    Returns
-    -------
-
     """
     # Test pymc and scipy distributions match for values and parameters
     # within the supported domain edges (excluding edges)
@@ -459,8 +462,8 @@ def check_logcdf(
         if decimal is None:
             decimal = select_by_precision(float64=6, float32=3)
 
-        for pt in product(domains, n_samples=n_samples):
-            params = dict(pt)
+        for point in product(domains, n_samples=n_samples):
+            params = dict(point)
             scipy_eval = scipy_logcdf(**params)
 
             value = params.pop("value")
@@ -496,7 +499,7 @@ def check_logcdf(
             for invalid_edge in invalid_edges:
                 if invalid_edge is not None:
                     test_params = valid_params.copy()  # Shallow copy should be okay
-                    test_params[invalid_param] = at.as_tensor_variable(invalid_edge)
+                    test_params[invalid_param] = pt.as_tensor_variable(invalid_edge)
                     # We need to remove `Assert`s introduced by checks like
                     # `assert_negative_support` and disable test values;
                     # otherwise, we won't be able to create the
@@ -559,8 +562,8 @@ def check_selfconsistency_discrete_logcdf(
     dist_logcdf = model.compile_fn(logcdf(rv, value))
     dist_logp = model.compile_fn(logp(rv, value))
 
-    for pt in product(domains, n_samples=n_samples):
-        params = dict(pt)
+    for point in product(domains, n_samples=n_samples):
+        params = dict(point)
         value = params.pop("value")
         values = np.arange(domain.lower, value + 1)
 
@@ -573,7 +576,7 @@ def check_selfconsistency_discrete_logcdf(
                 dist_logcdf({"value": value}),
                 sp.logsumexp([dist_logp({"value": value}) for value in values]),
                 decimal=decimal,
-                err_msg=str(pt),
+                err_msg=str(point),
             )
 
 
@@ -598,7 +601,7 @@ def assert_moment_is_expected(model, expected, check_finite_logp=True):
         logp_moment = (
             joint_logp(
                 (model["x"],),
-                rvs_to_values={model["x"]: at.constant(moment)},
+                rvs_to_values={model["x"]: pt.constant(moment)},
                 rvs_to_transforms={},
             )[0]
             .sum()
@@ -607,7 +610,7 @@ def assert_moment_is_expected(model, expected, check_finite_logp=True):
         assert np.isfinite(logp_moment)
 
 
-def pymc_random(
+def continuous_random_tester(
     dist,
     paramdomains,
     ref_rand,
@@ -629,12 +632,12 @@ def pymc_random(
     pymc_rand = compile_pymc([], model_dist)
 
     domains = paramdomains.copy()
-    for pt in product(domains, n_samples=100):
-        pt = pm.Point(pt, model=model)
-        pt.update(model_args)
+    for point in product(domains, n_samples=100):
+        point = pm.Point(point, model=model)
+        point.update(model_args)
 
         # Update the shared parameter variables in `param_vars`
-        for k, v in pt.items():
+        for k, v in point.items():
             nv = param_vars.get(k, model.named_vars.get(k))
             if nv.name in param_vars:
                 param_vars[nv.name].set_value(v)
@@ -645,13 +648,13 @@ def pymc_random(
         f = fails
         while p <= alpha and f > 0:
             s0 = pymc_rand()
-            s1 = floatX(ref_rand(size=size, **pt))
+            s1 = floatX(ref_rand(size=size, **point))
             _, p = st.ks_2samp(np.atleast_1d(s0).flatten(), np.atleast_1d(s1).flatten())
             f -= 1
-        assert p > alpha, str(pt)
+        assert p > alpha, str(point)
 
 
-def pymc_random_discrete(
+def discrete_random_tester(
     dist,
     paramdomains,
     valuedomain=None,
@@ -668,12 +671,12 @@ def pymc_random_discrete(
     pymc_rand = compile_pymc([], model_dist)
 
     domains = paramdomains.copy()
-    for pt in product(domains, n_samples=100):
-        pt = pm.Point(pt, model=model)
+    for point in product(domains, n_samples=100):
+        point = pm.Point(point, model=model)
         p = alpha
 
         # Update the shared parameter variables in `param_vars`
-        for k, v in pt.items():
+        for k, v in point.items():
             nv = param_vars.get(k, model.named_vars.get(k))
             if nv.name in param_vars:
                 param_vars[nv.name].set_value(v)
@@ -683,7 +686,7 @@ def pymc_random_discrete(
         f = fails
         while p <= alpha and f > 0:
             o = pymc_rand()
-            e = intX(ref_rand(size=size, **pt))
+            e = intX(ref_rand(size=size, **point))
             o = np.atleast_1d(o).flatten()
             e = np.atleast_1d(e).flatten()
             bins = min(20, max(len(set(e)), len(set(o))))
@@ -695,12 +698,29 @@ def pymc_random_discrete(
             else:
                 _, p = st.chisquare(observed + 1, expected + 1)
             f -= 1
-        assert p > alpha, str(pt)
+        assert p > alpha, str(point)
+
+
+class SeededTest:
+    random_seed = 20160911
+    random_state = None
+
+    @classmethod
+    def setup_class(cls):
+        nr.seed(cls.random_seed)
+
+    def setup_method(self):
+        nr.seed(self.random_seed)
+
+    def get_random_state(self, reset=False):
+        if self.random_state is None or reset:
+            self.random_state = nr.RandomState(self.random_seed)
+        return self.random_state
 
 
 class BaseTestDistributionRandom(SeededTest):
     """
-    This class provides a base for tests that new RandomVariables are correctly
+    Base class for tests that new RandomVariables are correctly
     implemented, and that the mapping of parameters between the PyMC
     Distribution and the respective RandomVariable is correct.
 
@@ -762,7 +782,7 @@ class BaseTestDistributionRandom(SeededTest):
     reference_dist: Optional[Callable] = None
     reference_dist_params: Optional[dict] = None
     expected_rv_op_params: Optional[dict] = None
-    checks_to_run = []
+    checks_to_run: List[str] = []
     size = 15
     decimal = select_by_precision(float64=6, float32=3)
 
@@ -864,3 +884,8 @@ def seeded_numpy_distribution_builder(dist_name: str) -> Callable:
     return lambda self: ft.partial(
         getattr(np.random.RandomState, dist_name), self.get_random_state()
     )
+
+
+def assert_no_rvs(var):
+    assert not any(isinstance(v.owner.op, RandomVariable) for v in ancestors([var]) if v.owner)
+    return var
