+++ tests/distributions/test_timeseries.py
@@ -44,8 +44,7 @@ from pymc.model import Model
 from pymc.pytensorf import floatX
 from pymc.sampling.forward import draw, sample_posterior_predictive
 from pymc.sampling.mcmc import sample
-from tests.distributions.util import assert_moment_is_expected
-from tests.helpers import select_by_precision
+from pymc.testing import assert_moment_is_expected, select_by_precision
 
 # Turn all warnings into errors for this module
 # Ignoring NumPy deprecation warning tracked in https://github.com/pymc-devs/pytensor/issues/146
