+++ tests/logprob/test_joint_logprob.py
@@ -58,7 +58,7 @@ import pymc as pm
 from pymc.logprob.abstract import logprob
 from pymc.logprob.joint_logprob import factorized_joint_logprob, joint_logp
 from pymc.logprob.utils import rvs_to_value_vars, walk_model
-from tests.helpers import assert_no_rvs
+from pymc.testing import assert_no_rvs
 from tests.logprob.utils import joint_logprob
