+++ tests/logprob/test_transforms.py
@@ -63,7 +63,7 @@ from pymc.logprob.transforms import (
     TransformValuesRewrite,
     transformed_variable,
 )
-from tests.helpers import assert_no_rvs
+from pymc.testing import assert_no_rvs
 from tests.logprob.utils import joint_logprob
