+++ tests/test_math.py
@@ -39,7 +39,8 @@ from pymc.math import (
     softmax,
 )
 from pymc.pytensorf import floatX
-from tests.helpers import SeededTest, verify_grad
+from pymc.testing import SeededTest
+from tests.helpers import verify_grad
 
 
 def test_kronecker():
