+++ tests/variational/test_minibatch_rv.py
@@ -20,8 +20,8 @@ from scipy import stats as st
 import pymc as pm
 
 from pymc import Normal, draw
+from pymc.testing import select_by_precision
 from pymc.variational.minibatch_rv import create_minibatch_rv
-from tests.helpers import select_by_precision
 from tests.test_data import gen1, gen2
